Move vehicle preference to RideGroupManager
